refactoring

+++ 1/src/on3d.py
@@ -5,9 +5,9 @@ from enum import Enum
 class Point():
     # точка с координатами (x, y, z)
     def __init__(self, *, x=0, y=0, z=0):
-        self.x: float = x
-        self.y: float = y
-        self.z: float = z
+        self.x: int = x
+        self.y: int = y
+        self.z: int = z
 
     def __str__(self) -> str:
         return f'({self.x},{self.y},{self.z})'
